+++ lib/src/dilithium.dart
@@ -15,6 +15,19 @@
  */
 part of dilithium;
 
+class Hive {
+
+  String name;
+  String src;
+  HttpRequest http;
+
+  Hive(this.name, this.src) {
+    this.http = new HttpRequest();
+
+  }
+
+}
+
 abstract class Dilithium extends Phaser.State {
 
   String path;
@@ -25,62 +38,106 @@ abstract class Dilithium extends Phaser.State {
    * Using resources...
    */
   static async.Future using(String path) {
+    bool isCocoon = window.navigator.appVersion.contains("CocoonJS");
 
-    // for now, we'll use javascript
-    var script = document.createElement('script');
-    // Secret Template Man
-    // Secret Template Man
-    // They've given you a browser, and taken away your brain.
-    script.setAttribute('src', 'packages/dilithium/liquid-0.0.7.min.js');
-    querySelector('head').append(script);
-
+    print("Start Using");
+    if (!isCocoon) {
+      // for now, we'll use javascript
+      var script = document.createElement('script');
+      // Secret Template Man
+      // Secret Template Man
+      // They've given you a browser, and taken away your brain.
+      script.setAttribute('src', 'packages/dilithium/liquid-0.0.7.min.js');
+      querySelector('head').append(script);
+    }
     async.Completer completer = new async.Completer();
 
     // get the browser locale setting
     String lang = window.navigator.language;
-    // simplify english to use default
-    if (lang == "en" || lang.startsWith("en-"))
-      lang = '';
-    else
-      lang = "-$lang";
 
-    //  get the core dilithium configuration
-    HttpRequest.getString("$path/config.yaml")
-    .then((String source) {
-      Li2Config config = new Li2Config(source, path);
-      String to_arrays = config.paths['arrays'];
-      String to_strings = config.paths['strings'];
-      String to_preferences = config.paths['preferences'];
+    /**
+     * get the dilithium configuration hive
+     */
+    HttpRequest httpConfig = new HttpRequest();
+    httpConfig.onReadyStateChange.listen((event) {
+      if (httpConfig.readyState == 4 && httpConfig.status == 200) {
 
-      HttpRequest.getString("$path/$to_preferences")
-      .then((String preferences) {
-        HttpRequest.getString("$path/$to_arrays")
-        .then((String arrays) {
-          // get localized strings (if they exist)
-          String to_locale = to_strings.replaceAll('.yaml', "-$lang.yaml");
-          HttpRequest.getString("$path/$to_locale")
-          .then((String strings) {
-            config.setStrings(strings);
-            config.setArrays(arrays);
-            config.setPreferences(preferences);
-            completer.complete(config);
-          })
-          // get default (if they don't exist)
-          .catchError((Error e) {
-            HttpRequest.getString("$path/$to_strings")
-            .then((String strings) {
-              config.setStrings(strings);
-              config.setArrays(arrays);
-              config.setPreferences(preferences);
+        /**
+         * Main Config loaded
+         */
+        Li2Config config = new Li2Config(httpConfig.responseText, path);
+        List hives = [];
+
+        /**
+         * Process hive sections, strings & arrays first
+         */
+        if (config.sections['strings'] != null) {
+          hives.add(new Hive('strings', config.sections['strings']));
+          if (config.locale) {
+            hives.add(new Hive('strings', config.sections['strings'].replaceAll('.yaml', "-$lang.yaml")));
+          }
+        }
+        if (config.sections['arrays'] != null) {
+          hives.add(new Hive('arrays', config.sections['strings']));
+        }
+        config.sections.forEach((name, src) {
+          if (name != 'strings' && name != 'arrays') {
+            hives.add(new Hive(name, src));
+          }
+        });
+        if (hives.length <= 0) {
+          completer.complete(config);
+          return;
+        }
+
+        /**
+         * Que them up
+         */
+        int flag = 0; // how to tell
+        int done = math.pow(2, hives.length)-1; // when we are done
+
+        hives.forEach((Hive hive) {
+          hive.http.onReadyStateChange.listen((event) {
+            for (int check=0; check<hives.length; check++) {
+              if (hives[check].http.readyState == 4
+              && (hives[check].http.status == 200 || hives[check].http.status == 404)) {
+                flag |= math.pow(2, check);
+              }
+            }
+
+            if (flag == done) {
+              /**
+               * When we are done, process them in order
+               */
+              hives.forEach((Hive section) {
+                if (section.http.status == 200) {
+                  if (section.http.responseText.length > 0)
+                    config.setSection(section.name, section.http.responseText);
+                }
+              });
               completer.complete(config);
-            });
+            }
           });
         });
-      });
-    });
-    return completer.future;
+        hives.forEach((Hive hive) {
+          hive.http.open('GET',"$path/${hive.src}");
+          hive.http.send();
+        });
 
+      } else if (httpConfig.readyState == 4 && httpConfig.status != 200) {
+        /**
+         * Nope, we got butkus
+         */
+        Li2Config config = new Li2Config('', path);
+        completer.complete(config);
+
+      }
+    });
+    httpConfig.open('GET', "$path/config.yaml");
+    httpConfig.send();
+    return completer.future;
   }
+
 
 
   /**