+++ changes.md
@@ -3,3 +3,9 @@
 * Added coffee version
 ### v0.2.1 - Nov 9, 2014
 * fix syncronous request should be async
+### v0.2.2 - Nov 19, 2014
+* Copy new config load from coffee to dart
+### v0.2.3 - Nov 20, 2014
+* templating now uses liquid.coffee
+* fix preload progress display
+* check for CocoonJS